S05P12A101-105 - [feat] : Todo 생성, 불러오기  / BE
- Todo 클래스 생성
- Todo 생성 API 구현
- Todo List 불러오기 API 구현

S05P12A101-105
S05P12A101-103

+++ backend/src/main/java/com/prossafy101/ohnaco/entity/todo/Category.java
@@ -10,7 +10,7 @@ import javax.persistence.Id;
 public class Category {
 
     @Id
-    private int categoryid;
+    private Long categoryid;
 
     private String categoryname;
 }
